src/requirements.txt
-i https://pypi.org/simple
bcrypt==4.1.1; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
cachecontrol==0.13.1; python_version >= '3.7'
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudevents==1.9.0
cryptography==41.0.7
deprecation==2.1.0
firebase-admin==6.2.0; python_version >= '3.7'
firebase-functions==0.1.2; python_version >= '3.10'
flask==3.0.0; python_version >= '3.8'
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-sqlalchemy==3.1.1; python_version >= '3.8'
functions-framework==3.5.0; python_version >= '3.5' and python_version < '4'
google-api-core[grpc]==2.14.0; platform_python_implementation != 'PyPy'
google-api-python-client==2.108.0; python_version >= '3.7'
google-auth==2.23.4; python_version >= '3.7'
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3; python_version >= '3.7'
google-cloud-firestore==2.13.1; python_version >= '3.7'
google-cloud-storage==2.13.0; python_version >= '3.7'
google-crc32c==1.5.0; python_version >= '3.7'
google-events==0.11.0; python_version >= '3.7'
google-resumable-media==2.6.0; python_version >= '3.7'
googleapis-common-protos==1.61.0; python_version >= '3.7'
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0; python_version >= '3.5'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.6; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
injector==0.21.0
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
marshmallow==3.20.1; python_version >= '3.8'
msgpack==1.0.7; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
proto-plus==1.22.3; python_version >= '3.6'
protobuf==4.25.1; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pyjwt[crypto]==2.8.0; python_version >= '3.7'
pyparsing==3.1.1; python_version >= '3.1'
pytest==7.4.3; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
sqlalchemy==2.0.23; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.1.0; python_version >= '3.8'
watchdog==3.0.0; python_version >= '3.7'
werkzeug==3.0.1; python_version >= '3.8'
